login feature refactored
