+++ src/components/App/component.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { BrowserRouter, Switch } from 'react-router-dom'
 import PrivateRoute from '../PrivateRoute'
-import PublicRoute from '../PublicRoute'
+import GuestRoute from '../GuestRoute'
 import LoginForm from '../LoginForm'
 import Dashboard from '../Dashboard'
 
@@ -9,8 +9,7 @@ import Dashboard from '../Dashboard'
 const App = () => (
   <BrowserRouter>
     <Switch>
-      <PublicRoute
-        restricted
+      <GuestRoute
         component={LoginForm}
         path="/login"
         exact
