+++ src/components/GuestRoute/index.js
@@ -1,0 +1,3 @@
+import GuestRoute from './container'
+
+export default GuestRoute
