--- src/components/PublicRoute/index.js
@@ -1,3 +1,0 @@
-import PublicRoute from './container'
-
-export default PublicRoute
